--- flyteplugins/go/tasks/plugins/k8s/ray/batchscheduler/config/config_test.go
@@ -1,15 +1,0 @@
-package config
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestNewConfig(t *testing.T) {
-	t.Run("New scheduler plugin config", func(t *testing.T) {
-		config := NewConfig()
-		assert.Equal(t, "", config.GetScheduler())
-		assert.Equal(t, "", config.GetParameters())
-	})
-}
